restructure annotate components
