+++ scm-ui/ui-components/src/index.ts
@@ -43,7 +43,6 @@ import {
 
 export { validation, urls, repositories };
 
-export { default as Annotate } from "./Annotate";
 export { default as DateFromNow } from "./DateFromNow";
 export { default as DateShort } from "./DateShort";
 export { default as ErrorNotification } from "./ErrorNotification";
