+++ scm-ui/ui-components/src/repos/annotate/AuthorImage.tsx
@@ -22,38 +22,13 @@
  * SOFTWARE.
  */
 
-export { Action } from "./Action";
-export { Link, Links, Collection, PagedCollection } from "./hal";
+import { AvatarImage } from "../../avatar";
+import styled from "styled-components";
 
-export { Me } from "./Me";
-export { DisplayedUser, User } from "./User";
-export { Group, Member } from "./Group";
+const AuthorImage = styled(AvatarImage)`
+  width: 1em;
+  height: 1em;
+  margin-right: 0.2em;
+`;
 
-export { Repository, RepositoryCollection, RepositoryGroup } from "./Repositories";
-export { RepositoryType, RepositoryTypeCollection } from "./RepositoryTypes";
-
-export { Branch, BranchRequest } from "./Branches";
-
-export { Person } from "./Person";
-
-export { Changeset, Contributor, ParentChangeset } from "./Changesets";
-
-export { AnnotatedSource, AnnotatedLine } from "./Annotate";
-
-export { Tag } from "./Tags";
-
-export { Config } from "./Config";
-
-export { IndexResources } from "./IndexResources";
-
-export { Permission, PermissionCreateEntry, PermissionCollection } from "./RepositoryPermissions";
-
-export { SubRepository, File } from "./Sources";
-
-export { SelectValue, AutocompleteObject } from "./Autocomplete";
-
-export { Plugin, PluginCollection, PluginGroup, PendingPlugins } from "./Plugin";
-
-export { RepositoryRole } from "./RepositoryRole";
-
-export { NamespaceStrategies } from "./NamespaceStrategies";
+export default AuthorImage;
