+++ scm-ui/ui-components/src/repos/annotate/actions.ts
@@ -22,38 +22,26 @@
  * SOFTWARE.
  */
 
-export { Action } from "./Action";
-export { Link, Links, Collection, PagedCollection } from "./hal";
+import { AnnotatedLine } from "@scm-manager/ui-types";
 
-export { Me } from "./Me";
-export { DisplayedUser, User } from "./User";
-export { Group, Member } from "./Group";
+type EnterLine = {
+  annotation: AnnotatedLine;
+  offset: number;
+  line: number;
+  type: "enter-line";
+};
 
-export { Repository, RepositoryCollection, RepositoryGroup } from "./Repositories";
-export { RepositoryType, RepositoryTypeCollection } from "./RepositoryTypes";
+type LeaveLine = {
+  line: number;
+  type: "leave-line";
+};
 
-export { Branch, BranchRequest } from "./Branches";
+type EnterPopover = {
+  type: "enter-popover";
+};
 
-export { Person } from "./Person";
+type LeavePopover = {
+  type: "leave-popover";
+};
 
-export { Changeset, Contributor, ParentChangeset } from "./Changesets";
-
-export { AnnotatedSource, AnnotatedLine } from "./Annotate";
-
-export { Tag } from "./Tags";
-
-export { Config } from "./Config";
-
-export { IndexResources } from "./IndexResources";
-
-export { Permission, PermissionCreateEntry, PermissionCollection } from "./RepositoryPermissions";
-
-export { SubRepository, File } from "./Sources";
-
-export { SelectValue, AutocompleteObject } from "./Autocomplete";
-
-export { Plugin, PluginCollection, PluginGroup, PendingPlugins } from "./Plugin";
-
-export { RepositoryRole } from "./RepositoryRole";
-
-export { NamespaceStrategies } from "./NamespaceStrategies";
+export type Action = EnterLine | LeaveLine | EnterPopover | LeavePopover;
