+++ scm-ui/ui-types/src/Annotate.ts
@@ -22,38 +22,18 @@
  * SOFTWARE.
  */
 
-export { Action } from "./Action";
-export { Link, Links, Collection, PagedCollection } from "./hal";
+import { Person } from "./Person";
 
-export { Me } from "./Me";
-export { DisplayedUser, User } from "./User";
-export { Group, Member } from "./Group";
+export type AnnotatedSource = {
+  lines: AnnotatedLine[];
+  language: string;
+};
 
-export { Repository, RepositoryCollection, RepositoryGroup } from "./Repositories";
-export { RepositoryType, RepositoryTypeCollection } from "./RepositoryTypes";
-
-export { Branch, BranchRequest } from "./Branches";
-
-export { Person } from "./Person";
-
-export { Changeset, Contributor, ParentChangeset } from "./Changesets";
-
-export { AnnotatedSource, AnnotatedLine } from "./Annotate";
-
-export { Tag } from "./Tags";
-
-export { Config } from "./Config";
-
-export { IndexResources } from "./IndexResources";
-
-export { Permission, PermissionCreateEntry, PermissionCollection } from "./RepositoryPermissions";
-
-export { SubRepository, File } from "./Sources";
-
-export { SelectValue, AutocompleteObject } from "./Autocomplete";
-
-export { Plugin, PluginCollection, PluginGroup, PendingPlugins } from "./Plugin";
-
-export { RepositoryRole } from "./RepositoryRole";
-
-export { NamespaceStrategies } from "./NamespaceStrategies";
+export type AnnotatedLine = {
+  author: Person;
+  code: string;
+  description: string;
+  lineNumber: number;
+  revision: string;
+  when: Date;
+};
