+++ scm-ui/ui-types/src/Changesets.ts
@@ -25,11 +25,7 @@
 import { Collection, Links } from "./hal";
 import { Tag } from "./Tags";
 import { Branch } from "./Branches";
-
-export type Person = {
-  name: string;
-  mail?: string;
-};
+import { Person } from "./Person";
 
 export type Changeset = Collection & {
   id: string;
